Unified draw calls, attempt 1
Replaced all GLPoly calls with a batching procedure, to try to group
all meshes and use a final GL call to draw them all in one step,
instead of many.
Lava appears wrong with this, like it has a forced
"gl1_overbrightbits 1" to it. Do I use colorArray to fix this?
Also, need to put a limit for the vertices and indexes counter.