VERY WIP implementation of triple lightmap
We have 3 copies of the same lightmap in memory at all times, so we
can alternate between them for each new frame. This should help with
speed on tile rendering platforms, since changing the same texture
that is being rendered (or was recently rendered) is a very
expensive operation there.

We still need this to coexist with traditional lightmaps.